indexserver: factor out index logic
This makes it possible to swap out zoekt-archive-index for a clone based
approach.
